requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
discord-components==0.0.0.1
discord.py==2.4.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
numpy==1.26.3
pandas==2.2.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.32.3
six==1.16.0
tbb==2021.12.0
tzdata==2023.4
urllib3==2.2.2
yarl==1.9.4
apscheduler==3.9.1
asyncpg==0.27.0
python-dotenv